+++ scene.c
@@ -15,467 +15,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <unistd.h>
 
-#include <libft.h>
 #include "cub3d.h"
 #include "scene.h"
-
-int	c3_scene_read_int(t_c3_scene_parser *buf)
-{
-	int		result;
-	char	ch;
-
-	ch = buf->getc(buf->container);
-	result = 0;
-	while (ch >= '0' && ch <= '9')
-	{
-		result = result * 10 + (ch - '0');
-		ch = buf->getc(buf->container);
-	}
-	if (ch >= 0)
-		buf->ungetc(buf->container);
-	return (result);
-}
-
-t_c3_token	c3_scene_get_token_expect(
-	t_c3_scene_parser *buf, char next_char, t_c3_token token)
-{
-	int	ch;
-
-	ch = buf->getc(buf->container);
-	if (ch == next_char)
-		return (token);
-	buf->ungetc(buf->container);
-	return (C3_SCENE_TOKEN_UNKNOWN);
-}
-
-t_c3_token	c3_scene_get_token_s(t_c3_scene_parser *buf)
-{
-	int	ch;
-
-	ch = buf->getc(buf->container);
-	if (ch == 'O')
-		return (C3_SCENE_TOKEN_SO);
-	buf->ungetc(buf->container);
-	if (ch == ' ')
-		return (C3_SCENE_TOKEN_S);
-	return (C3_SCENE_TOKEN_UNKNOWN);
-}
-
-t_c3_token	c3_scene_get_token_head(t_c3_scene_parser *buf, int ch)
-{
-	buf->is_beginning_of_line = 0;
-	if (ch < 0)
-		return (C3_SCENE_TOKEN_EOF);
-	if (ch == 'R')
-		return (C3_SCENE_TOKEN_R);
-	if (ch == 'N')
-		return (c3_scene_get_token_expect(buf, 'O', C3_SCENE_TOKEN_NO));
-	if (ch == 'S')
-		return (c3_scene_get_token_s(buf));
-	if (ch == 'W')
-		return (c3_scene_get_token_expect(buf, 'E', C3_SCENE_TOKEN_WE));
-	if (ch == 'E')
-		return (c3_scene_get_token_expect(buf, 'A', C3_SCENE_TOKEN_EA));
-	if (ch == 'F')
-		return (C3_SCENE_TOKEN_F);
-	if (ch == 'C')
-		return (C3_SCENE_TOKEN_C);
-	if (ch >= 0)
-		buf->ungetc(buf->container);
-	else
-		return (C3_SCENE_TOKEN_EOF);
-	return (C3_SCENE_TOKEN_POSSIBLY_MAP);
-}
-
-t_c3_token	c3_scene_get_token(t_c3_scene_parser *buf)
-{
-	int	ch;
-
-	ch = buf->getc(buf->container);
-	while (ch == '\n')
-	{
-		buf->is_beginning_of_line = 1;
-		ch = buf->getc(buf->container);
-	}
-	if (buf->is_beginning_of_line)
-		return (c3_scene_get_token_head(buf, ch));
-	while (ch == ' ' || ch == '\t')
-		ch = buf->getc(buf->container);
-	if (ch >= '0' && ch <= '9')
-	{
-		buf->ungetc(buf->container);
-		buf->int_value = c3_scene_read_int(buf);
-		return (C3_SCENE_TOKEN_NUM);
-	}
-	if (ch == ',')
-		return (C3_SCENE_TOKEN_COMMA);
-	if (ch < 0)
-		return (C3_SCENE_TOKEN_EOF);
-	return (C3_SCENE_TOKEN_UNKNOWN);
-}
-
-int			c3_scene_get_int(t_c3_scene_parser *buf)
-{
-	return (buf->int_value);
-}
-
-const char	*c3_scene_get_string(t_c3_scene_parser *buf)
-{
-	char	ch;
-	int		index;
-
-	ch = buf->getc(buf->container);
-	while (ch == ' ' || ch == '\t')
-		ch = buf->getc(buf->container);
-	index = 0;
-	while (ch > 0 && ch != '\n' && index < C3_STRING_BUFFER_SIZE - 1)
-	{
-		buf->string_value[index++] = ch;
-		ch = buf->getc(buf->container);
-	}
-	buf->ungetc(buf->container);
-	buf->string_value[index] = '\0';
-	return (buf->string_value);
-}
-
-const char	*c3_scene_get_rest(t_c3_scene_parser *buf)
-{
-	char	ch;
-	int		index;
-
-	ch = buf->getc(buf->container);
-	index = 0;
-	while (ch > 0 && ch != '\n' && index < C3_STRING_BUFFER_SIZE - 1)
-	{
-		buf->string_value[index++] = ch;
-		ch = buf->getc(buf->container);
-	}
-	if (ch >= 0)
-		buf->ungetc(buf->container);
-	buf->string_value[index] = '\0';
-	return (buf->string_value);
-}
-
-void	c3_strbuf_ungetc(t_c3_scene_container cont)
-{
-	t_c3_strbuf	*buf;
-
-	buf = cont.strbuf;
-	if (buf->index > 0)
-		buf->index--;
-}
-
-int		c3_strbuf_getc(t_c3_scene_container cont)
-{
-	t_c3_strbuf	*buf;
-	char		ch;
-
-	buf = cont.strbuf;
-	ch = buf->str[buf->index];
-	if (ch)
-	{
-		buf->index++;
-		return (ch);
-	}
-	return (-1);
-}
-
-void	c3_file_ungetc(t_c3_scene_container cont)
-{
-	t_c3_file	*file;
-
-	file = cont.file;
-	if (file->is_ungotten)
-	{
-		c3_log("ungetc: too many chars ungotten.");
-		return ;
-	}
-	file->is_ungotten = 1;
-	file->ungotten = file->last_char;
-}
-
-int		c3_file_getc(t_c3_scene_container cont)
-{
-	t_c3_file	*file;
-	char		buf[1];
-	int			read_result;
-
-	file = cont.file;
-	if (file->is_ungotten)
-	{
-		file->is_ungotten = 0;
-		file->last_char = file->ungotten;
-		return (file->ungotten);
-	}
-	read_result = read(file->fd, buf, 1);
-	if (read_result == 0)
-		return (-1);
-	if (read_result < 0)
-	{
-		perror("read failed");
-		return (-1);
-	}
-	file->last_char = buf[0];
-	return (buf[0]);
-}
-
-t_c3_parse_result	c3_scene_parse_resolution(
-	t_c3_scene *scene, t_c3_scene_parser *buf)
-{
-	int			width;
-	int			height;
-
-	if (c3_scene_get_token(buf) == C3_SCENE_TOKEN_NUM)
-	{
-		width = c3_scene_get_int(buf);
-		if (c3_scene_get_token(buf) == C3_SCENE_TOKEN_NUM)
-		{
-			height = c3_scene_get_int(buf);
-			scene->resolution.x = width;
-			scene->resolution.y = height;
-			return (C3_PARSE_SUCCESS);
-		}
-	}
-	buf->error = "Invalid resolution format";
-	return (C3_PARSE_FAIL);
-}
-
-t_c3_parse_result	c3_scene_parse_texture(
-	t_c3_scene *scene, t_c3_object_type typ, t_c3_scene_parser *buf)
-{
-	const char	*path;
-
-	path = c3_scene_get_string(buf);
-	scene->tex_path[typ] = ft_strdup(path);
-	if (!scene->tex_path[typ])
-	{
-		buf->error = strerror(errno);
-		return (C3_PARSE_FAIL);
-	}
-	return (C3_PARSE_SUCCESS);
-}
-
-static t_c3_parse_result	parse_color(
-	unsigned int *result, t_c3_scene_parser *buf)
-{
-	unsigned int	r;
-	unsigned int	g;
-	unsigned int	b;
-
-	if (c3_scene_get_token(buf) == C3_SCENE_TOKEN_NUM)
-	{
-		r = c3_scene_get_int(buf);
-		if (c3_scene_get_token(buf) == C3_SCENE_TOKEN_COMMA
-			&& c3_scene_get_token(buf) == C3_SCENE_TOKEN_NUM)
-		{
-			g = c3_scene_get_int(buf);
-			if (c3_scene_get_token(buf) == C3_SCENE_TOKEN_COMMA
-				&& c3_scene_get_token(buf) == C3_SCENE_TOKEN_NUM)
-			{
-				b = c3_scene_get_int(buf);
-				*result = ((r & 0xff) << 16) + ((g & 0xff) << 8) + (b & 0xff);
-				return (C3_PARSE_SUCCESS);
-			}
-		}
-	}
-	buf->error = "Invalid color format";
-	return (C3_PARSE_FAIL);
-}
-
-t_c3_parse_result	c3_scene_parse_floor(
-	t_c3_scene *scene, t_c3_scene_parser *buf)
-{
-	return (parse_color(&scene->color_floor, buf));
-}
-
-t_c3_parse_result	c3_scene_parse_ceiling(
-	t_c3_scene *scene, t_c3_scene_parser *buf)
-{
-	return (parse_color(&scene->color_ceiling, buf));
-}
-
-void	c3_map_rows_cleanup(t_c3_map_rows *rows)
-{
-	if (rows)
-	{
-		free(rows->row);
-		c3_map_rows_cleanup(rows->next);
-		free(rows);
-	}
-}
-
-t_c3_map_rows	*t_c3_map_rows_create(
-	t_c3_scene_parser *buf, t_c3_map_rows *rows, const char *str)
-{
-	t_c3_map_rows	*new_row;
-
-	new_row = malloc(sizeof(t_c3_map_rows));
-	if (!new_row)
-	{
-		buf->error = strerror(errno);
-		c3_map_rows_cleanup(rows);
-		return (NULL);
-	}
-	new_row->next = rows;
-	new_row->row = ft_strdup(str);
-	return (new_row);
-}
-
-t_c3_map_rows	*c3_scene_parse_read_rows(
-	t_c3_scene *scene, t_c3_scene_parser *buf)
-{
-	t_c3_token		tok;
-	t_c3_map_rows	*rows;
-	int				width;
-
-	rows = NULL;
-	tok = C3_SCENE_TOKEN_POSSIBLY_MAP;
-	while (tok == C3_SCENE_TOKEN_POSSIBLY_MAP)
-	{
-		if (!(rows = t_c3_map_rows_create(buf, rows, c3_scene_get_rest(buf))))
-			return (NULL);
-		width = ft_strlen(rows->row);
-		tok = c3_scene_get_token(buf);
-		scene->map_width = scene->map_width < width ? width : scene->map_width;
-		scene->map_height++;
-	}
-	if (tok != C3_SCENE_TOKEN_EOF)
-	{
-		buf->error = "Invalid map format";
-		c3_map_rows_cleanup(rows);
-		return (NULL);
-	}
-	return (rows);
-}
-
-void				c3_scene_populate_map(
-	t_c3_scene *scene, t_c3_map_rows *rows)
-{
-	t_c3_map_rows	*new_row;
-	int				num_rows;
-	int				width;
-	char			*dest;
-
-	num_rows = scene->map_height;
-	while (num_rows > 0)
-	{
-		dest = &scene->map[(num_rows - 1) * scene->map_width];
-		width = ft_strlen(rows->row);
-		ft_memcpy(dest, rows->row, width);
-		ft_memset(dest + width, ' ', scene->map_width - width);
-		free(rows->row);
-		new_row = rows->next;
-		free(rows);
-		rows = new_row;
-		num_rows--;
-	}
-}
-
-t_c3_parse_result	c3_scene_parse_map(
-	t_c3_scene *scene, t_c3_scene_parser *buf)
-{
-	t_c3_map_rows	*rows;
-
-	if (!(rows = c3_scene_parse_read_rows(scene, buf)))
-		return (C3_PARSE_FAIL);
-	if (!(scene->map = malloc(scene->map_width * scene->map_height)))
-	{
-		buf->error = strerror(errno);
-		return (C3_PARSE_FAIL);
-	}
-	c3_scene_populate_map(scene, rows);
-	return (C3_PARSE_SUCCESS);
-}
-
-int					c3_scene_try_wall(
-	t_c3_scene *scene, t_c3_scene_parser *buf,
-	t_c3_token tok, t_c3_parse_result *result)
-{
-	unsigned int		i;
-
-	i = C3_SCENE_TOKEN_NO;
-	while (i < C3_SCENE_TOKEN_NO + 5)
-	{
-		if (tok == i)
-		{
-			if (buf->is_specified[i])
-			{
-				buf->error = "some textures specified multiple times";
-				*result = C3_PARSE_FAIL;
-				return (0);
-			}
-			buf->is_specified[i] = 1;
-			*result = c3_scene_parse_texture(
-				scene, i - (C3_SCENE_TOKEN_NO - C3_OBJTYPE_WALL_N), buf);
-			if (*result != C3_PARSE_SUCCESS)
-				return (0);
-		}
-		i++;
-	}
-	return (1);
-}
-
-int					c3_scene_try_ceiling(
-	t_c3_scene *scene, t_c3_scene_parser *buf,
-	t_c3_token tok, t_c3_parse_result *result)
-{
-	if (tok == C3_SCENE_TOKEN_C)
-	{
-		if (buf->is_specified[C3_SCENE_TOKEN_C])
-		{
-			buf->error = "C specified multiple times";
-			*result = C3_PARSE_FAIL;
-			return (0);
-		}
-		buf->is_specified[C3_SCENE_TOKEN_C] = 1;
-		*result = c3_scene_parse_ceiling(scene, buf);
-		if (*result != C3_PARSE_SUCCESS)
-			return (0);
-	}
-	return (1);
-}
-
-int					c3_scene_try_floor(
-	t_c3_scene *scene, t_c3_scene_parser *buf,
-	t_c3_token tok, t_c3_parse_result *result)
-{
-	if (tok == C3_SCENE_TOKEN_F)
-	{
-		if (buf->is_specified[C3_SCENE_TOKEN_F])
-		{
-			buf->error = "F specified multiple times";
-			*result = C3_PARSE_FAIL;
-			return (0);
-		}
-		buf->is_specified[C3_SCENE_TOKEN_F] = 1;
-		*result = c3_scene_parse_floor(scene, buf);
-		if (*result != C3_PARSE_SUCCESS)
-			return (0);
-	}
-	return (1);
-}
-
-int					c3_scene_try_resolution(
-	t_c3_scene *scene, t_c3_scene_parser *buf,
-	t_c3_token tok, t_c3_parse_result *result)
-{
-	if (tok == C3_SCENE_TOKEN_R)
-	{
-		if (buf->is_specified[C3_SCENE_TOKEN_R])
-		{
-			buf->error = "R specified multiple times";
-			*result = C3_PARSE_FAIL;
-			return (0);
-		}
-		buf->is_specified[C3_SCENE_TOKEN_R] = 1;
-		*result = c3_scene_parse_resolution(scene, buf);
-		if (*result != C3_PARSE_SUCCESS)
-			return (0);
-	}
-	return (1);
-}
+#include "scene_int.h"
 
 t_c3_parse_result	c3_scene_parse(t_c3_scene *scene, t_c3_scene_parser *buf)
 {
@@ -504,7 +47,7 @@ t_c3_parse_result	c3_scene_parse(t_c3_scene *scene, t_c3_scene_parser *buf)
 	return (result);
 }
 
-void	c3_scene_init(t_c3_scene *scene)
+void				c3_scene_init(t_c3_scene *scene)
 {
 	int	i;
 
@@ -516,7 +59,7 @@ void	c3_scene_init(t_c3_scene *scene)
 	scene->map = NULL;
 }
 
-void	c3_scene_cleanup(t_c3_scene *scene)
+void				c3_scene_cleanup(t_c3_scene *scene)
 {
 	int				i;
 
@@ -530,7 +73,8 @@ void	c3_scene_cleanup(t_c3_scene *scene)
 	free(scene->map);
 }
 
-int		c3_scene_parser_init_with_file(t_c3_scene_parser *buf, const char *path)
+int					c3_scene_parser_init_with_file(
+	t_c3_scene_parser *buf, const char *path)
 {
 	int	fd;
 	int	i;
